+++ transforms/code/malware/Dockerfile.python
@@ -17,7 +17,7 @@ FROM docker.io/clamav/clamav:latest AS clamav
 FROM base AS clamav-local
 USER root
 COPY --chown=dpk:root --from=clamav /var/lib/clamav/ /var/lib/clamav/
-COPY --chown=dpk:root clamd.conf /etc/clamav/clamd.conf
+COPY --chown=dpk:root clamd.python.conf /etc/clamav/clamd.conf
 RUN freshclam \
     && mkdir -p /var/run/clamav \
     && chown -R dpk:root /var/run/clamav /var/log/clamav /var/lib/clamav
@@ -26,13 +26,17 @@ USER dpk
 
 FROM base AS malware
 
+COPY --chown=dpk:root --from=clamav-local /var/lib/clamav/ /var/lib/clamav/
+COPY --chown=dpk:root --from=clamav-local /etc/clamav/clamd.conf /etc/clamav/clamd.conf
+COPY --chown=dpk:root --from=clamav-local /var/log/clamav/clamav.log /var/log/clamav/clamav.log
+COPY --chown=dpk:root --from=clamav-local /var/run/clamav /var/run/clamav
+
 RUN pip install --upgrade --no-cache-dir pip 
 
 # install pytest
 RUN pip install --no-cache-dir pytest
 
 # Create a user and use it to run the transform
-RUN useradd -ms /bin/bash dpk
 USER dpk
 WORKDIR /home/dpk
 ARG DPK_WHEEL_FILE_NAME
