+++ transforms/code/malware/Dockerfile.ray
@@ -1,15 +1,44 @@
 ARG BASE_IMAGE=docker.io/rayproject/ray:2.24.0-py310
-FROM ${BASE_IMAGE}
+
+FROM ${BASE_IMAGE} AS base
 
 # see https://docs.openshift.com/container-platform/4.17/openshift_images/create-images.html#use-uid_create-images
 USER root
 RUN chown ray:root /home/ray && chmod g=u /home/ray
 USER ray
 
-RUN pip install --upgrade --no-cache-dir pip 
+RUN pip install --upgrade --no-cache-dir pip
 
-# install pytest
 RUN pip install --no-cache-dir pytest
+
+USER root
+RUN apt -y update \
+    && apt install -y clamav-daemon \
+    && apt-get clean \
+    && rm -rf /var/lib/apt/lists/*
+USER ray
+
+FROM docker.io/clamav/clamav:latest AS clamav
+FROM base AS clamav-local
+USER root
+COPY --chown=ray:0 --from=clamav /var/lib/clamav/ /var/lib/clamav/
+COPY --chown=ray:0 clamd.ray.conf /etc/clamav/clamd.conf
+RUN freshclam \
+    && mkdir -p /var/run/clamav \
+    && chown -R ray:0 /var/run/clamav /var/log/clamav /var/lib/clamav \
+    && chmod -R 777 /var/run/clamav /var/log/clamav /var/lib/clamav
+USER ray
+CMD ["/bin/bash", "-c", "clamd --debug --foreground"]
+
+FROM base AS malware
+
+ARG EXTRA_ARTIFACTORY_URL
+
+COPY --from=clamav-local --chown=ray:0 /var/lib/clamav/ /var/lib/clamav/
+COPY --from=clamav-local --chown=ray:0 /etc/clamav/clamd.conf /etc/clamav/clamd.conf
+COPY --from=clamav-local --chown=ray:0 /var/log/clamav /var/log/clamav
+COPY --from=clamav-local --chown=ray:0 /var/run/clamav /var/run/clamav
+
 ARG DPK_WHEEL_FILE_NAME
 ARG TRANSFORM_NAME
 
@@ -25,6 +54,15 @@ RUN pip install --no-cache-dir -r requirements.txt
 
 # Set environment
 ENV PYTHONPATH /home/ray
+
+USER root
+RUN rm -rf /var/log/clamav && mkdir -p /var/log/clamav && chown -R ray:0 /var/log/clamav
+USER ray
+RUN chmod -R 777 /var/run/clamav /var/log/clamav
+
+
+COPY --chown=ray:root test/ test/
+COPY --chown=ray:root test-data/ test-data/
 
 # Put these at the end since they seem to upset the docker cache.
 ARG BUILD_DATE
