+++ transforms/code/malware/clamd.ray.conf
