+++ transforms/code/malware/dpk_malware/local.py
@@ -14,7 +14,7 @@ import os
 
 from data_processing.data_access import DataAccessLocal
 from data_processing.utils import get_logger
-from malware_transform import MalwareTransform, check_clamd
+from dpk_malware.transform import MalwareTransform, check_clamd
 
 
 logger = get_logger(__name__)
