+++ transforms/code/malware/dpk_malware/local_python.py
@@ -16,8 +16,8 @@ from pathlib import Path
 
 from data_processing.runtime.pure_python import PythonTransformLauncher
 from data_processing.utils import ParamsUtils, get_logger
-from malware_transform import check_clamd
-from transforms.code.malware.dpk_malware.transform_python import MalwarePythonTransformConfiguration
+from dpk_malware.transform import check_clamd
+from dpk_malware.runtime import MalwarePythonTransformConfiguration
 
 
 logger = get_logger(__name__)
