+++ transforms/code/malware/dpk_malware/ray/local.py
@@ -16,8 +16,8 @@ from pathlib import Path
 
 from data_processing.utils import ParamsUtils, get_logger
 from data_processing_ray.runtime.ray import RayTransformLauncher
-from malware_transform import check_clamd
-from malware_transform_ray import MalwareRayTransformConfiguration
+from dpk_malware.transform import check_clamd
+from dpk_malware.ray.runtime import MalwareRayTransformConfiguration
 
 
 logger = get_logger(__name__)
