+++ transforms/code/malware/dpk_malware/ray/runtime.py
@@ -14,7 +14,7 @@ from data_processing_ray.runtime.ray import (
     RayTransformLauncher,
     RayTransformRuntimeConfiguration,
 )
-from malware_transform import MalwareTransformConfiguration
+from dpk_malware.transform import MalwareTransformConfiguration
 
 
 logger = get_logger(__name__)
