+++ transforms/code/malware/kfp_ray/malware_wf.py
@@ -27,7 +27,7 @@ from workflow_support.compile_utils import (
 S3_SECRET = "s3-secret"
 
 # the name of the job script
-EXEC_SCRIPT_NAME: str = "malware_transform_ray.py"
+EXEC_SCRIPT_NAME: str = "-m dpk_malware.ray.runtime"
 
 task_image = "quay.io/dataprep1/data-prep-kit/malware-ray:latest"
