--- transforms/code/malware/python/.dockerignore
@@ -1,1 +1,0 @@
-venv/
