--- transforms/code/malware/ray/.dockerignore
@@ -1,1 +1,0 @@
-venv/
