+++ transforms/code/malware/requirements.txt
@@ -1,0 +1,1 @@
+clamd==1.0.2
