+++ transforms/code/malware/test/test_malware.py
@@ -16,7 +16,7 @@ import pyarrow as pa
 from data_processing.test_support.transform.table_transform_test import (
     AbstractTableTransformTest,
 )
-from malware_transform import MalwareTransform
+from dpk_malware.transform import MalwareTransform
 
 
 table = pa.Table.from_pydict(
