+++ transforms/code/malware/test/test_malware_python.py
@@ -19,8 +19,8 @@ from data_processing.runtime.pure_python import PythonTransformLauncher
 from data_processing.test_support.launch.transform_test import (
     AbstractTransformLauncherTest,
 )
-from malware_transform import INPUT_COLUMN_KEY, OUTPUT_COLUMN_KEY
-from malware_transform_python import MalwarePythonTransformConfiguration
+from dpk_malware.transform import INPUT_COLUMN_KEY, OUTPUT_COLUMN_KEY
+from dpk_malware.runtime import MalwarePythonTransformConfiguration
 
 
 class TestPythonNOOPTransform(AbstractTransformLauncherTest):
