+++ transforms/code/malware/test/test_malware_ray.py
@@ -19,8 +19,8 @@ from data_processing.test_support.launch.transform_test import (
     AbstractTransformLauncherTest,
 )
 from data_processing_ray.runtime.ray import RayTransformLauncher
-from malware_transform import INPUT_COLUMN_KEY, OUTPUT_COLUMN_KEY
-from malware_transform_ray import MalwareRayTransformConfiguration
+from dpk_malware.transform import INPUT_COLUMN_KEY, OUTPUT_COLUMN_KEY
+from dpk_malware.ray.runtime import MalwareRayTransformConfiguration
 
 
 class TestRayNOOPTransform(AbstractTransformLauncherTest):
